Move currentLocale to multipaz library and make it a property.
Also fix compilation error with Swift TestApp.

Test: Compose TestApp and Swift TestApp compiles and runs.
