--- multipaz-doctypes/src/androidMain/kotlin/org/multipaz/doctypes/localization/NativeLocale.android.kt
@@ -1,9 +1,0 @@
-package org.multipaz.doctypes.localization
-
-import java.util.Locale
-
-internal actual object NativeLocale {
-    actual fun currentLocale(): String {
-        return Locale.getDefault().language
-    }
-}
