--- multipaz-doctypes/src/commonMain/kotlin/org/multipaz/doctypes/localization/NativeLocale.kt
@@ -1,5 +1,0 @@
-package org.multipaz.doctypes.localization
-
-internal expect object NativeLocale {
-     fun currentLocale(): String
-}
