--- multipaz-doctypes/src/wasmJsMain/kotlin/org/multipaz/doctypes/localization/NativeLocale.wasmJs.kt
@@ -1,9 +1,0 @@
-package org.multipaz.doctypes.localization
-
-import kotlinx.browser.window
-
-internal actual object NativeLocale {
-    actual fun currentLocale(): String {
-        return window.navigator.language.substringBefore("-")
-    }
-}
