+++ multipaz/src/commonMain/kotlin/org/multipaz/util/LocaleUtil.kt
@@ -1,0 +1,6 @@
+package org.multipaz.util
+
+/**
+ * The current locale as a BCP-47 string.
+ */
+expect val currentLocale: String
