+++ multipaz/src/javaSharedMain/kotlin/org/multipaz/util/LocaleUtil.jvm.kt
@@ -1,0 +1,8 @@
+package org.multipaz.util
+
+import java.util.Locale
+
+actual val currentLocale: String
+    get() {
+        return Locale.getDefault().language
+    }
