+++ multipaz/src/jsMain/kotlin/org/multipaz/util/LocaleUtil.js.kt
@@ -1,13 +1,10 @@
-package org.multipaz.doctypes.localization
+package org.multipaz.util
 
-import kotlin.js.js
-
-internal actual object NativeLocale {
-    actual fun currentLocale(): String {
+actual val currentLocale: String
+    get() {
         return try {
             js("navigator.language || 'en'").toString().substringBefore("-")
         } catch (e: Throwable) {
             "en"
         }
     }
-}
