+++ multipaz/src/wasmJsMain/kotlin/org/multipaz/util/LocaleUtil.wasmJs.kt
@@ -1,0 +1,8 @@
+package org.multipaz.util
+
+import kotlinx.browser.window
+
+actual val currentLocale: String
+    get() {
+        return window.navigator.language.substringBefore("-")
+    }
